Polish apache/incubator-dubbo#3946 : Refactor Dubbo metadata service
