+++ dubbo-metadata/pom.xml
@@ -32,26 +32,12 @@
     <description>The metadata module of Dubbo project</description>
 
     <dependencies>
+
         <dependency>
             <groupId>org.apache.dubbo</groupId>
-            <artifactId>dubbo-config-api</artifactId>
+            <artifactId>dubbo-rpc-api</artifactId>
             <version>${revision}</version>
             <optional>true</optional>
-        </dependency>
-
-        <!-- Test dependencies -->
-        <dependency>
-            <groupId>org.apache.dubbo</groupId>
-            <artifactId>dubbo-rpc-dubbo</artifactId>
-            <version>${revision}</version>
-            <scope>test</scope>
-        </dependency>
-
-        <dependency>
-            <groupId>org.apache.dubbo</groupId>
-            <artifactId>dubbo-remoting-netty4</artifactId>
-            <version>${revision}</version>
-            <scope>test</scope>
         </dependency>
 
     </dependencies>
