--- dubbo-metadata/src/main/resources/META-INF/dubbo/internal/org.apache.dubbo.metadata.export.MetadataServiceExporter
@@ -1,1 +1,0 @@
-default=org.apache.dubbo.metadata.export.ConfigurableMetadataServiceExporter
