--- dubbo-metadata/src/test/resources/META-INF/dubbo/internal/org.apache.dubbo.registry.RegistryFactory
@@ -1,1 +1,0 @@
-mock=org.apache.dubbo.metadata.export.MockRegistryFactory
